% Generated by build script
\name{eudata_dataset}
\alias{eudata_dataset}
\title{Get details for a specific EU dataset}
\usage{
eudata_dataset()
}
\arguments{
\item{id}{Dataset identifier (e.g. "groenklimaatassen-gent")}
}
\value{tibble with one row: id, title, description, country, catalog,}
\description{Get details for a specific EU dataset}
